+++ internal/provider/names.go
@@ -528,4 +528,5 @@ const (
 	keyYearlySchedule                               = "yearly_schedule"
 	keyYearStartMonth                               = "year_start_month"
 	keyZone                                         = "zone"
+	keyZones                                        = "zones"
 )