+++ docs/guides/changelog.md
@@ -39,6 +39,9 @@ page_title: "Changelog"
 * New data source added for `rubrik_gcp_service_accounts` which returns the GCP service accounts RSC has
   discovered for a cloud account, for use with the `rubrik_gcp_cloud_cluster` resource.
   [[docs](../data-sources/gcp_service_accounts.md)]
+* New data source added for `rubrik_gcp_regions` which returns the GCP regions and their availability zones
+  RSC supports for a cloud account, for use with the `rubrik_gcp_cloud_cluster` resource.
+  [[docs](../data-sources/gcp_regions.md)]
 
 ## v1.9.0
 * **Breaking Change:** When the `CNP_AZURE_SQL_SLA_REVAMP` feature is enabled for the account, Azure SQL Database and